Edited v3
Implemented inputTime variable so user has a limited amount of time before all inputs are read. inputTime decrements each time score increases by 10.
Not sure that we need to do interrupts, but this needs to be tested. I think we might be fine without them
Added conditional case where if any incorrect input is sensed that the user is told their final score and sent to an infinite while loop -- need to decide how we want to reset and start the game